requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autobahn==20.2.2
Automat==20.2.0
autopep8==1.5
cffi==1.14.0
channels==2.4.0
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
dj-database-url==0.5.0
Django==3.0.3
django-heroku==0.3.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.19
PyHamcrest==2.0.1
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
pyOpenSSL==19.1.0
pytz==2019.3
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
Twisted==19.10.0
txaio==20.1.1
typed-ast==1.4.1
whitenoise==5.0.1
wrapt==1.11.2
zope.interface==4.7.1
